+++ packages/plugins/typescript/msw/tests/msw.spec.ts
@@ -73,8 +73,8 @@ describe('msw', () => {
 
     const result = await plugin(null, documents, {});
 
-    expect(result.content).toContain(`const { ${variables.join(', ')} } = req.variables`);
-    expect(result.content).toContain(`ctx.data({ ${selection} })`);
+    expect(result.content).toContain(`const { ${variables.join(', ')} } = variables`);
+    expect(result.content).toContain(`data: { ${selection} }`);
     expect(result.content).toMatchSnapshot(
       'content with variables and selection JSDoc documentation',
     );
@@ -87,14 +87,16 @@ describe('msw', () => {
     // handler variable and result type
     const queryVariablesType = `${importOperationTypesFrom}.${queryName}QueryVariables`;
     const queryType = `${importOperationTypesFrom}.${queryName}Query`;
-    expect(result.content).toContain(`GraphQLRequest<${queryVariablesType}>`);
-    expect(result.content).toContain(`GraphQLContext<${queryType}>`);
+    expect(result.content).toContain(
+      `GraphQLResponseResolver<${queryType}, ${queryVariablesType}>`,
+    );
     expect(result.content).toContain(`graphql.query<${queryType}, ${queryVariablesType}>`);
 
     const mutationVariablesType = `${importOperationTypesFrom}.${mutationName}MutationVariables`;
     const mutationType = `${importOperationTypesFrom}.${mutationName}Mutation`;
-    expect(result.content).toContain(`GraphQLRequest<${mutationVariablesType}>`);
-    expect(result.content).toContain(`GraphQLContext<${mutationType}>`);
+    expect(result.content).toContain(
+      `GraphQLResponseResolver<${mutationType}, ${mutationVariablesType}>`,
+    );
     expect(result.content).toContain(`graphql.mutation<${mutationType}, ${mutationVariablesType}>`);
 
     expect(result.content).toMatchSnapshot(