+++ packages/plugins/typescript/react-apollo/tests/react-apollo.spec.ts
@@ -70,6 +70,16 @@ describe('React Apollo', () => {
     }
   `);
 
+  const fragmentDoc = parse(/* GraphQL */ `
+    fragment RepositoryFields on Repository {
+      full_name
+      html_url
+      owner {
+        avatar_url
+      }
+    }
+  `);
+
   const validateTypeScript = async (
     output: Types.PluginOutput,
     testSchema: GraphQLSchema,
@@ -1286,6 +1296,75 @@ export function useSubmitRepositoryMutation(baseOptions?: Apollo.MutationHookOpt
       await validateTypeScript(content, schema, docs, {});
     });
 
+    it.each([
+      {
+        document: `query Feed { feed { ...Feed } } `,
+        requiredVariables: false,
+      },
+      {
+        document: `mutation Feed { feed { ...Feed } } `,
+        requiredVariables: false,
+      },
+      {
+        document: `query Feed($something: Boolean) { feed { ...Feed } } `,
+        requiredVariables: false,
+      },
+      {
+        document: `query Feed($something: Boolean!) { feed { ...Feed } } `,
+        requiredVariables: true,
+      },
+      {
+        document: `mutation Feed($something: Boolean!) { feed { ...Feed } } `,
+        requiredVariables: false,
+      },
+      {
+        document: `query Feed($something: Boolean!, $somethingElse: Boolean!) { feed { ...Feed } } `,
+        requiredVariables: true,
+      },
+      {
+        document: `query Feed($something: Boolean, $somethingElse: Boolean!) { feed { ...Feed } } `,
+        requiredVariables: true,
+      },
+      {
+        document: `query Feed($something: Boolean! = true) { feed { ...Feed } } `,
+        requiredVariables: false,
+      },
+      {
+        document: `query Feed($something: Boolean! = true, $somethingElse: Boolean) { feed { ...Feed } } `,
+        requiredVariables: false,
+      },
+    ])(
+      'Should have variables as required if query contains required variables',
+      async ({ document, requiredVariables }) => {
+        const docs = [
+          {
+            location: '',
+            document: parse(document),
+          },
+        ];
+
+        const result = (await plugin(
+          schema,
+          docs,
+          {},
+          {
+            outputFile: 'graphql.tsx',
+          },
+        )) as Types.ComplexPluginOutput;
+
+        const requiredVariableString =
+          ' & ({ variables: FeedQueryVariables; skip?: boolean; } | { skip: boolean; }) ';
+
+        if (requiredVariables) {
+          expect(result.content).toContain(requiredVariableString);
+        } else {
+          expect(result.content).not.toContain(requiredVariableString);
+        }
+
+        await validateTypeScript(result, schema, docs, {});
+      },
+    );
+
     it('Should generate deduped hooks for query and mutation', async () => {
       const documents = parse(/* GraphQL */ `
         query FeedQuery {
@@ -1434,6 +1513,10 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
 
       expect(content.content).toBeSimilarStringTo(`
       export type FeedLazyQueryHookResult = ReturnType<typeof useFeedLazyQuery>;
+      `);
+
+      expect(content.content).toBeSimilarStringTo(`
+      export type FeedSuspenseQueryHookResult = ReturnType<typeof useFeedSuspenseQuery>;
       `);
 
       expect(content.content).toBeSimilarStringTo(`
@@ -1719,6 +1802,40 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       }`);
       await validateTypeScript(content, schema, docs, {});
     });
+    it('should generate suspense query hooks', async () => {
+      const documents = parse(/* GraphQL */ `
+        query feed {
+          feed {
+            id
+            commentCount
+            repository {
+              full_name
+              html_url
+              owner {
+                avatar_url
+              }
+            }
+          }
+        }
+      `);
+      const docs = [{ location: '', document: documents }];
+
+      const content = (await plugin(
+        schema,
+        docs,
+        { withHooks: true, withComponent: false, withHOC: false },
+        {
+          outputFile: 'graphql.tsx',
+        },
+      )) as Types.ComplexPluginOutput;
+
+      expect(content.content).toBeSimilarStringTo(`
+      export function useFeedSuspenseQuery(baseOptions?: Apollo.SuspenseQueryHookOptions<FeedQuery, FeedQueryVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useSuspenseQuery<FeedQuery, FeedQueryVariables>(FeedDocument, options);
+      }`);
+      await validateTypeScript(content, schema, docs, {});
+    });
     it('should generate lazy query hooks with proper hooksSuffix', async () => {
       const documents = parse(/* GraphQL */ `
         query feed {
@@ -1750,6 +1867,40 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       export function useFeedLazyQueryMySuffix(baseOptions?: Apollo.LazyQueryHookOptions<FeedQuery, FeedQueryVariables>) {
         const options = {...defaultOptions, ...baseOptions}
         return Apollo.useLazyQuery<FeedQuery, FeedQueryVariables>(FeedDocument, options);
+      }`);
+      await validateTypeScript(content, schema, docs, {});
+    });
+    it('should generate suspense query hooks with proper hooksSuffix', async () => {
+      const documents = parse(/* GraphQL */ `
+        query feed {
+          feed {
+            id
+            commentCount
+            repository {
+              full_name
+              html_url
+              owner {
+                avatar_url
+              }
+            }
+          }
+        }
+      `);
+      const docs = [{ location: '', document: documents }];
+
+      const content = (await plugin(
+        schema,
+        docs,
+        { withHooks: true, withComponent: false, withHOC: false, hooksSuffix: 'MySuffix' },
+        {
+          outputFile: 'graphql.tsx',
+        },
+      )) as Types.ComplexPluginOutput;
+
+      expect(content.content).toBeSimilarStringTo(`
+      export function useFeedSuspenseQueryMySuffix(baseOptions?: Apollo.SuspenseQueryHookOptions<FeedQuery, FeedQueryVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useSuspenseQuery<FeedQuery, FeedQueryVariables>(FeedDocument, options);
       }`);
       await validateTypeScript(content, schema, docs, {});
     });
@@ -2176,7 +2327,7 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       await validateTypeScript(content, schema, docs, {});
     });
 
-    it('should import Operations from one external file and use it in useQuery and useLazyQuery', async () => {
+    it('should import Operations from one external file and use it in useQuery, useLazyQuery and useSuspenseQuery', async () => {
       const config: ReactApolloRawPluginConfig = {
         documentMode: DocumentMode.external,
         importDocumentNodeExternallyFrom: 'path/to/documents',
@@ -2202,6 +2353,12 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       export function useTestLazyQuery(baseOptions?: Apollo.LazyQueryHookOptions<TestQuery, TestQueryVariables>) {
         const options = {...defaultOptions, ...baseOptions}
         return Apollo.useLazyQuery<TestQuery, TestQueryVariables>(Operations.test, options);
+      }
+      `);
+      expect(content.content).toBeSimilarStringTo(`
+      export function useTestSuspenseQuery(baseOptions?: Apollo.SuspenseQueryHookOptions<TestQuery, TestQueryVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useSuspenseQuery<TestQuery, TestQueryVariables>(Operations.test, options);
       }
       `);
       await validateTypeScript(content, schema, docs, {});
@@ -2457,6 +2614,12 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       }
       `);
       expect(content.content).toBeSimilarStringTo(`
+      export function useTestOneSuspenseQuery(baseOptions?: Apollo.SuspenseQueryHookOptions<TestOneQuery, TestOneQueryVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useSuspenseQuery<TestOneQuery, TestOneQueryVariables>(Operations.testOne, options);
+      }
+      `);
+      expect(content.content).toBeSimilarStringTo(`
       export function useTestTwoMutation(baseOptions?: Apollo.MutationHookOptions<TestTwoMutation, TestTwoMutationVariables>) {
         const options = {...defaultOptions, ...baseOptions}
         return Apollo.useMutation<TestTwoMutation, TestTwoMutationVariables>(Operations.testTwo, options);
@@ -2550,7 +2713,7 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       await validateTypeScript(content, schema, docs, {});
     });
 
-    it('should import Operations from near operation file for useQuery and useLazyQuery', async () => {
+    it('should import Operations from near operation file for useQuery, useLazyQuery and useSuspenseQuery', async () => {
       const config: ReactApolloRawPluginConfig = {
         documentMode: DocumentMode.external,
         importDocumentNodeExternallyFrom: 'near-operation-file',
@@ -2576,6 +2739,12 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       export function useTestLazyQuery(baseOptions?: Apollo.LazyQueryHookOptions<TestQuery, TestQueryVariables>) {
         const options = {...defaultOptions, ...baseOptions}
         return Apollo.useLazyQuery<TestQuery, TestQueryVariables>(Operations.test, options);
+      }
+      `);
+      expect(content.content).toBeSimilarStringTo(`
+      export function useTestSuspenseQuery(baseOptions?: Apollo.SuspenseQueryHookOptions<TestQuery, TestQueryVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useSuspenseQuery<TestQuery, TestQueryVariables>(Operations.test, options);
       }
       `);
       await validateTypeScript(content, schema, docs, {});
@@ -2632,6 +2801,40 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       export const TestComponent = (props: TestComponentProps) => (
         <ApolloReactComponents.Mutation<TestMutation, TestMutationVariables> mutation={Operations.test} {...props} />
       );`);
+      await validateTypeScript(content, schema, docs, {});
+    });
+
+    it('should import fragments from near operation file for useFragment', async () => {
+      const config: ReactApolloRawPluginConfig = {
+        documentMode: DocumentMode.external,
+        importDocumentNodeExternallyFrom: 'near-operation-file',
+        withComponent: false,
+        withHooks: true,
+        withHOC: false,
+        withFragmentHooks: true,
+      };
+
+      const docs = [{ location: 'path/to/document.graphql', document: fragmentDoc }];
+
+      const content = (await plugin(schema, docs, config, {
+        outputFile: 'graphql.tsx',
+      })) as Types.ComplexPluginOutput;
+
+      expect(content.prepend[0]).toBeSimilarStringTo(`import * as Apollo from '@apollo/client';`);
+
+      expect(content.content).toBeSimilarStringTo(`
+      export function useRepositoryFieldsFragment<F = { id: string }>(identifiers: F) {
+        return Apollo.useFragment<RepositoryFieldsFragment>({
+          fragment: RepositoryFieldsFragmentDoc,
+          fragmentName: "RepositoryFields",
+          from: {
+            __typename: "Repository",
+            ...identifiers,
+          },
+        });
+      }
+      export type RepositoryFieldsFragmentHookResult = ReturnType<typeof useRepositoryFieldsFragment>;
+      `);
       await validateTypeScript(content, schema, docs, {});
     });
 
@@ -2826,6 +3029,12 @@ export function useListenToCommentsSubscription(baseOptions?: Apollo.Subscriptio
       export function useTestOneLazyQuery(baseOptions?: Apollo.LazyQueryHookOptions<TestOneQuery, TestOneQueryVariables>) {
         const options = {...defaultOptions, ...baseOptions}
         return Apollo.useLazyQuery<TestOneQuery, TestOneQueryVariables>(Operations.testOne, options);
+      }
+      `);
+      expect(content.content).toBeSimilarStringTo(`
+      export function useTestOneSuspenseQuery(baseOptions?: Apollo.SuspenseQueryHookOptions<TestOneQuery, TestOneQueryVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useSuspenseQuery<TestOneQuery, TestOneQueryVariables>(Operations.testOne, options);
       }
       `);
       expect(content.content).toBeSimilarStringTo(`