+++ packages/plugins/typescript/react-query/CHANGELOG.md
@@ -1,6 +1,21 @@
 # @graphql-codegen/typescript-react-query
 
-## 5.1.0
+## 6.1.0
+
+### Minor Changes
+
+- [#552](https://github.com/dotansimha/graphql-code-generator-community/pull/552)
+  [`b403529`](https://github.com/dotansimha/graphql-code-generator-community/commit/b403529bb2805c58285707e4615532140c7e4156)
+  Thanks [@neil585456525](https://github.com/neil585456525)! - Fix importOperationTypesFrom missing
+  prefix
+
+## 6.0.0
+
+### Major Changes
+
+- [#434](https://github.com/dotansimha/graphql-code-generator-community/pull/434)
+  [`935b51f07`](https://github.com/dotansimha/graphql-code-generator-community/commit/935b51f0777047102cc1c33a1a18a4527902e0f9)
+  Thanks [@neil585456525](https://github.com/neil585456525)! - Support react-query v5
 
 ### Minor Changes