+++ packages/plugins/typescript/vue-apollo/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @graphql-codegen/typescript-vue-apollo
 
+## 4.1.1
+
+### Patch Changes
+
+- [#557](https://github.com/dotansimha/graphql-code-generator-community/pull/557)
+  [`c2b4a1b`](https://github.com/dotansimha/graphql-code-generator-community/commit/c2b4a1b6f13808745860c814ae38667cfee8e0b9)
+  Thanks [@FloEdelmann](https://github.com/FloEdelmann)! - Make lazy query's variables optional
+
+## 4.1.0
+
+### Minor Changes
+
+- [#363](https://github.com/dotansimha/graphql-code-generator-community/pull/363)
+  [`8e829d54c`](https://github.com/dotansimha/graphql-code-generator-community/commit/8e829d54c4017ee03184795395911894424dc0c4)
+  Thanks [@swunderlich](https://github.com/swunderlich)! - Possibility to add a clientId for the use
+  of multiple apollo clients
+
 ## 4.0.0
 
 ### Major Changes